#7 영화 컴포넌트 분리
